Nodes now check for taps

+++ Circl/app/src/main/java/com/calvinpelletier/circl/Coord.java
@@ -5,10 +5,10 @@ package com.calvinpelletier.circl;
  */
 public class Coord {
 
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public Coord(int x, int y) {
+    public Coord(float x, float y) {
         this.x = x;
         this.y = y;
     }
